support dirent related system calls

+++ libc/sysdeps/mos/common/sys/types.h
@@ -25,7 +25,7 @@
 
 
 /*** Required in GCC compilation **/
-typedef int ino_t;
+typedef unsigned long ino_t;
 typedef int dev_t;
 typedef long off_t;
 /********************************/
